simplify the bip44 derivation path scheme (even for the chimeric one)
